Extract web span handling
